perf(data store): separate Datastore API
